+++ src/core/tests/policy_tests.rs
@@ -122,3 +122,29 @@ fn parse_boolean_expression_additional_tests() {
         & ap("why not", "here too");
     assert_eq!(expected_ap, access_policy);
 }
+
+#[test]
+fn verify_access_policy() {
+    let access_policy = "(Department::HR || Department::R&D) && Level::level 2";
+    let policy = Policy::new(1000)
+        .add_axis(
+            "Level",
+            &["level 1", "level 2", "level 3", "level 4", "level 5"],
+            true,
+        )
+        .unwrap()
+        .add_axis("Department", &["R&D", "HR", "MKG", "fin"], false)
+        .unwrap();
+    AccessPolicy::verify_access_policy(access_policy, &policy).unwrap();
+
+    assert!(AccessPolicy::verify_access_policy(
+        "(Department1111::HR || Department::R&D) && Level::level 2",
+        &policy
+    )
+    .is_err());
+    assert!(AccessPolicy::verify_access_policy(
+        "(Department::HR111111 || Department::R&D) && Level::level 2",
+        &policy
+    )
+    .is_err());
+}